+++ src/helpers/html.ts
@@ -1,4 +1,4 @@
-import { EventBus, getLocale, InterpolateFunction, textUtil, TimeRange } from '@grafana/data';
+import { DataFrame, EventBus, getLocale, InterpolateFunction, PanelData, textUtil, TimeRange } from '@grafana/data';
 import { config, locationService } from '@grafana/runtime';
 import { TimeZone } from '@grafana/schema';
 import Handlebars from 'handlebars';
@@ -26,6 +26,8 @@ export const generateHtml = ({
   replaceVariables,
   eventBus,
   options,
+  panelData,
+  dataFrame,
 }: {
   data: Record<string, unknown>;
   content: string;
@@ -35,6 +37,8 @@ export const generateHtml = ({
   replaceVariables: InterpolateFunction;
   eventBus: EventBus;
   options: PanelOptions;
+  panelData: PanelData;
+  dataFrame?: DataFrame;
 }): { html: string; unsubscribe?: unknown } => {
   /**
    * Variable
@@ -59,6 +63,8 @@ export const generateHtml = ({
       'replaceVariables',
       'locationService',
       'eventBus',
+      'panelData',
+      'dataFrame',
       helpers
     );
     unsubscribe = func(
@@ -70,6 +76,8 @@ export const generateHtml = ({
       replaceVariables,
       locationService,
       eventBus,
+      panelData,
+      dataFrame,
       helpers
     );
   }